requirements.txt
djangorestframework
django
psycopg2-binary 


